Move cars to separate directory, add parking_slot to the program
