+++ cars/not_steerable_car.py
@@ -1,0 +1,5 @@
+from cars.car import Car
+
+class NotSteerableCar(Car):
+  def update(self, dt):
+    return
